% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloudrr.R
\name{wordcloudrr}
\alias{wordcloudrr}
\title{Generate javascript wordcloud}
\usage{
wordcloudrr(words, freq, cols = "#000000", shape = "circle", dir = NULL,
  width = 800, height = 800)
}
\arguments{
\item{words}{character vector of words to be plotted}

\item{freq}{numeric vector of word frequency}

\item{cols}{character vector of HTML RGB codes}

\item{shape}{Shape of wordcloud, "circle", "triangle", "star"}

\item{dir}{character, path to export the wordcloud, NULL is temp. dir}

\item{width}{numeric, width of the wordcloud}

\item{height}{numeric, height of the wordcloud}
}
\value{
Nothing
}
\description{
A drop-in replacement for wordcloud() to generate javascript wordcloud.
}
\examples{
\dontrun{
words <- c("hello", 'ngramrr', 'chainsawriot', 'weiborr', 'hkgovrr', 'xx')
freq <- c(150, 80, 50,100, 20, 10)
cols <- c("#FF0000", "#FF0000", "#FF0000", "#00FF00", "#00FF00", "#0000FF")
wordcloudrr(words, freq, cols)
}
}
\author{
Chung-hong Chan <chainsawtiney@gmail.com>, with wordcloud2.js written by timedream and TaffyDB.
}

